rename ToolsFragment to ToolsListFragment
