+++ app/src/main/res/layout/tools_fragment.xml
@@ -8,7 +8,7 @@
         <import type="androidx.lifecycle.LiveData" />
         <import type="java.util.List" />
 
-        <variable name="fragment" type="org.cru.godtools.ui.tools.ToolsFragment" />
+        <variable name="fragment" type="org.cru.godtools.ui.tools.ToolsListFragment" />
         <variable name="tools" type="LiveData&lt;List&lt;org.cru.godtools.model.Tool&gt;&gt;" />
     </data>
